restored default skin, adjusted font colors, removed some compiler warnings
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code@274 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CXAnimationPlayer.cpp
@@ -319,7 +319,8 @@ void CXAnimationPlayer::addFacesToBuffer(s32 meshbuffernr, CXFileReader::SXMesh&
 	{
 		video::S3DVertex* vertex = &buf->Vertices[0];
 		fcnt = buf->Indices.size();
-		for (s32 u=0; u<fcnt; u+=3)
+
+		for (s32 u=0; u<(s32)fcnt; u+=3)
 		{
 			core::plane3d<f32> p(vertex[buf->Indices[u+0]].Pos,
 				vertex[buf->Indices[u+1]].Pos, vertex[buf->Indices[u+2]].Pos);
